separated kubernetes and helm packages
