+++ compliance-test-kit/BDD-and-TDD/features/UC002-F001-Expose-APIs-Create-API-Resource.feature
@@ -9,17 +9,17 @@ Feature: UC002-F001 Expose APIs: Create API Resource
     Scenario: Create API Resource for Core API
         Given An example package 'productcatalog-v1beta2' with a 'productcatalog' component with '1' API in its 'coreFunction' segment
         When I install the 'productcatalog-v1beta2' package
-        Then I should see the 'productcatalogmanagement' API resource
+        Then I should see the 'productcatalogmanagement' API resource on the 'productcatalog' component
 
     Scenario: Create API Resource for Management API
         Given An example package 'productcatalog-v1beta2' with a 'productcatalog' component with '1' API in its 'managementFunction' segment
         When I install the 'productcatalog-v1beta2' package
-        Then I should see the 'metrics' API resource
+        Then I should see the 'metrics' API resource on the 'productcatalog' component
 
     Scenario: Create API Resource for Security API
         Given An example package 'productcatalog-v1beta2' with a 'productcatalog' component with '1' API in its 'securityFunction' segment
         When I install the 'productcatalog-v1beta2' package
-        Then I should see the 'partyrole' API resource
+        Then I should see the 'partyrole' API resource on the 'productcatalog' component
