+++ compliance-test-kit/BDD-and-TDD/features/UC002-F003-Expose-APIs-Verify-API-implementation-is-ready.feature
@@ -9,16 +9,16 @@ Feature: UC002-F003 Expose APIs: Verify API implementation is ready
     Scenario: Verify API Resource is ready for Core API
         Given An example package 'productcatalog-v1beta2' with a 'productcatalog' component with '1' API in its 'coreFunction' segment
         When I install the 'productcatalog-v1beta2' package
-        Then I should see the 'productcatalogmanagement' API resource with an implementation ready status on the Service Mesh or Gateway
+        Then I should see the 'productcatalogmanagement' API resource on the 'productcatalog' component with an implementation ready status on the Service Mesh or Gateway
 
     Scenario: Verify API Resource is ready for Management API
         Given An example package 'productcatalog-v1beta2' with a 'productcatalog' component with '1' API in its 'managementFunction' segment
         When I install the 'productcatalog-v1beta2' package
-        Then I should see the 'metrics' API resource with an implementation ready status on the Service Mesh or Gateway
+        Then I should see the 'metrics' API resource on the 'productcatalog' component with an implementation ready status on the Service Mesh or Gateway
 
     Scenario: Verify API Resource is ready for Security API
         Given An example package 'productcatalog-v1beta2' with a 'productcatalog' component with '1' API in its 'securityFunction' segment
         When I install the 'productcatalog-v1beta2' package
-        Then I should see the 'partyrole' API resource with an implementation ready status on the Service Mesh or Gateway
+        Then I should see the 'partyrole' API resource on the 'productcatalog' component with an implementation ready status on the Service Mesh or Gateway
