+++ compliance-test-kit/BDD-and-TDD/package.json
@@ -13,7 +13,8 @@
     "@cucumber/cucumber": "^7.3.2",
     "chai": "^4.3.6",
     "chai-http": "^4.3.0",
-    "kubernetes-utils": "file:kubernetes-utils",
+    "resource-inventory-utils-kubernetes": "file:resource-inventory-utils-kubernetes",
+    "package-manager-utils-helm": "file:package-manager-utils-helm",
     "yaml": "^2.1.3"
   }
 }
